컴포넌트화 후 정상 작동, 그러나 css 실시간 수정 불가
